move back extractor to TicTacToe file
