+++ camel/environments/tic_tac_toe.py
@@ -13,11 +13,36 @@
 # ========= Copyright 2023-2024 @ CAMEL-AI.org. All Rights Reserved. =========
 import math
 import random
+import re
 from typing import Any, ClassVar, Dict, List, Literal, Optional, Tuple
 
 from camel.environments.models import Action, Observation
 from camel.environments.multi_step import MultiStepEnv
-from camel.extractors import BaseExtractor, TicTacToeExtractorStrategy
+from camel.extractors import BaseExtractor, BaseExtractorStrategy
+
+
+class MoveExtractor(BaseExtractorStrategy):
+    r"""A strategy for extracting Tic Tac Toe actions from text."""
+
+    async def extract(self, text: str) -> Optional[str]:
+        r"""Extract a valid Tic Tac Toe move from text.
+
+        Looks for a pattern '<Action> n' where n is a digit between 1 and 9.
+
+        Args:
+            text (str): The text to extract the action from.
+
+        Returns:
+            Optional[str]: The extracted move as a string, or None if no valid
+                move is found.
+        """
+        match = re.search(r"<Action>\s*(\d+)", text)
+        if match:
+            move = match.group(1)
+            # Validate that the move is in range 1-9
+            if move.isdigit() and 1 <= int(move) <= 9:
+                return move
+        return None
 
 
 class Opponent:
@@ -188,7 +213,7 @@ class TicTacToeEnv(MultiStepEnv):
         Args:
             extractor (Optional[BaseExtractor]): Extractor to process LLM
                 responses. If None, a default extractor with
-                TicTacToeExtractorStrategy will be used. (default: :obj:`None`)
+                MoveExtractor will be used. (default: :obj:`None`)
             max_steps (Optional[int]): Maximum steps per episode.
                 (default: :obj:`None`)
             play_style (Literal["optimal", "random"]): The strategy for the
@@ -197,9 +222,7 @@ class TicTacToeEnv(MultiStepEnv):
             **kwargs: Additional environment parameters.
         """
         if extractor is None:
-            extractor = BaseExtractor(
-                pipeline=[[TicTacToeExtractorStrategy()]]
-            )
+            extractor = BaseExtractor(pipeline=[[MoveExtractor()]])
         super().__init__(extractor, max_steps, **kwargs)
         self.opponent = Opponent(play_style=play_style)
