+++ camel/extractors/__init__.py
@@ -18,7 +18,6 @@ from .python_strategies import (
     PythonListStrategy,
     PythonSetStrategy,
     PythonTupleStrategy,
-    TicTacToeExtractorStrategy,
 )
 
 __all__ = [
@@ -29,5 +28,4 @@ __all__ = [
     "PythonDictStrategy",
     "PythonSetStrategy",
     "PythonTupleStrategy",
-    "TicTacToeExtractorStrategy",
 ]
